Added: join, merge, merge_hash
previous `join` to join a sequence to a string is renamed to `str_join`.
`join` is like a database join. You provide another sequence and it
builds the cartesian product, but only for those elements that match
$predicate.

`join` creates a tuple of both entries. With `merge` and `merge_hash`
you can specify a way to get back a "flat" structure.