Removed optional functional calling style for wrap, concat
I tested it, but in Perl its normal usage to call "static" methods
with "->" instead of "::". Those statics only can be called with "->".

I consider it is okay when they cannot be called as a method. They
are not supposed to work this way.

But you can call every chaining function as a standalone
call with `::` instead of using `->`.