$default must be explicitly set in min*, max*, first and last functions
Previously $default was an optional argument, when not provided it default
to undef. I decided to remove that ability. A user must provide a
default value when the sequence is empty. This leads to correct
programs. Users of the API must think what happens when there will be
returned `undef`.

If they really want to return `undef`, because it is different to other
values, then they must explicitly pass `undef` as the default.

Yes, it is annoying, and it should be annoying.