Extended from_sub
Throughout the system so far. It is potentially possible to ask
for the next value, even after an $it returned undef. At the moment
this was not happening, otherwise tests would probably fail.

But to ensure that $it does not return a value after the first undef,
an explicit $abort flag is introduced. This has an additional benefit.

The $it can be freed after the first undef. Also other objects can go
out of scope. And for example filehandles and so on are properly closed
as soon as possible.