splitted fold into two disctinct functions
Now there is fold and fold_mut. The automatic detection of a reftype
will not proberly work. When the sequence contains a sequence of objects
then it would always expect to mutate the state, even if that is not
wanted or possible. Could also be an immutable object.

Instead of DWIM that sometimes can be wrong, we specify with fold and
fold_mut what we want. Less confusion and we can still do what we want.