Reworked group_by & group_fold
group_by maps a sequence by a $key function. Instead of returning an
hash of arrays it is now a sequence again, containing the 'Key and
the grouped entries again as a sequence.

This way further processing of the grouped entries is possible.

group_fold does the same, but allows the user to define a folding function
to combine the values.