Changed do, doi
do() is for side-effects. Because of this, it was coded to consume
the whole sequence at once, and do something with every element. But,
this behaviour is already available with iter() and iteri().

do() is now changed to really behave like a sequence operation. It does
not do any side-effects in advance. It only does the side-effects
for those elements that get requested.

For example doing `$seq->do(...)->take(5)` will now only trigger
the side-effect for five elements even when they are more elements in
the sequence. And only at the moment they are requested.

This opens up new abilities. For example printing/logging elements from
a sequence. While the side-effect is really triggered only in the exact
moment an element is requested from the sequence.

See: ./examples/AoC-2023/01p2.pl