Changed:new,bless  Added:get,extract,set,push  Removed:add
They are not aliases anymore. bless expects an hashref and new expects
an even-sized lists. They idea to be "intelligent" and be intelligent
when just one argument is passed turned out to be failed.

Maybe not with Hash->new() but by implementing Array->new() with the
same logic it turned out to be a mess. So now new() just expects
its values the same as Array->new() just expects its values and both
construct an Hash/Array from it.

If otherwise the user hash an hash-ref or array-ref he uses Hash->bless()
or Array->bless().

Otherwise added function to get and set values. Removed `add` and favoring
`set` instead. `extract` for returning multiple values at once.

`push` for often needed - push an entry to a key that should be an array.