Changed:group_by and methods returning Hash
group_by should really return a Hash instead of a Seq. group_by anyway
has to go through the whole data-structure once. After a grouping
operatin you probably want to access data by its 'Key. Otherwise you
probably don't need to group the data. So a Hash is the best representation
for it. The return value of Hash of Array also shows that a Sequence
must be traversed completely and is not lazy.

Other methods returning a hash now returns blessed Hash.