Changed Hash::choose
Usually i expect to explicitly provide an $default value to avoid
undef, or at least a user must explicitly pass undef as a value when
he wants some.

But in the case of choose it makes less sense. The user anyway must return
an undef in the lambda he passes, so makes it obvious that an undef can
be returned. And it also makes sense that choose itself returns an
"optional". And i use undef for that case anyway.