Hash::push
Pushing into a key that was not an array made the code throw an error.
I didn't handle that case before and expected the user to know about
how the data-structure is formed.

But actually it makes sense that whenever a value is pushed onto a key
that this value just becomes the first value of the array.

Now also adds Array blessings to an array.