Trying to eliminate undef
In Sq i try to eliminate the presence of undef as much as possible.
Usually undef is threated like an optional. So it makes sense
that push/unshift also does some kind of undef checking and aborts
on an undef. This behaviour is the same when we create a new Array
with Array->new().

This does not eliminate the chance of undef completely as someone can
use the built-in push. If someone really want to push undefs
into an array he can do it. But this way it also has some
advantages of using ->push and ->unshift methods instead of using
the built-in functions provided by perl.