Renamed count to length
methods named "count" and "count_by" should count something and
return the result as a hash. Because of this getting the amount
of a collection should be renamed to "length" to avoid this collision.

But this counting function was already available as Array::as_hash so
i renamed that function to Array::count. Array::as_hash is now a function
turning an array into a hash.

I guess this is a better naming and cleanup