Inlining from_sub() call in Seq::from_array
from_sub() has a certain overhead because of function calls, keeping
memory references and so on. Inlining code makes it faster. I would
wish a feature like F# inline. Neverthless one some functions that
are crucial and very often used, or work as a basic to providing data
this inlining can be done manually to provide some speed improvements.