Performance improvement (40%)
The filtering code for List data-structures in benchmarks/filtering.pl
runs about 40% faster just by inlining the head(),tail() functions
in iter(). It shows how much of a performance penalty you can get
by just using function calls.

Perl is also a pre-compiled language, whenever you start a script it gets
completely parsed and compiled into an opcode before execution. But it
is intended to be fast to execute, so it doesn't have a lot of ahead of
pre-compile time optimizations like other languages.

This has one disadvantage as someone who programs with Perl has to know
the performance cost of certain things, what also is an advantage.

In other languages this is the same. Let's pick F# as an example. I could
for example mark the function `head` and `tail` as `inline` and give the
compiler the hint that he should always inline these function call so
no function call appear in the compiled code.

This let's you allow write more better high-level code, better understandable
for humans and still with the advantage that it produces fast "low-level"
code at the cost of maybe some longer compilation time.

This is basically also what compile-time macros in LISP languages do.