Change: Option::fold
Changed the order in which the anonymous subroutine receives its
parameter. Usually the call to fold looks like this.

$opt->fold($state, $f)

or

Option::fold($opt, $state, $f)

as you can see in both cases you provide the $opt and then $state. And this
is also the order in which the parameters are passed to `$f`.