Changed internal representation of Option
Instead of an array with two values where the first value represents
if we have the Some or None case i just use an empty array or an array
with one value.

Instead of checking the first value i just check the length of the array
to determine if we have a value or not. This should be faster and consumes
less memory.

Even though that in first benchmarks i don't see a difference in speed
I keep that design. Also makes code a little bit shorter.