Changed fold
Up to this point in all `fold` functions the $f_state functions was called
with the argument order ($state,$x). Swapped the order to ($x,$state).
This means the first argument is always one element of a sequence,array,
list and so on, then the $state is passed.

The reason for this change is that this order makes more sense in how the
code is written. For example

    $seq->fold($state, $sub)
    Seq::fold($seq, $state, $sub)

as you can see you always write `$seq` and then `$state` in both function
calls. And this is also the order in which `$sub` gets the argument passed.

I previously changed Option::fold to also this order as it makes more
sense and it is easier to remember.