Seq: Changed: min,max, min_by,max_by
`min` and `max` is changed to return an optional instead of expecting a $default
argument.

`min_by` and `max_by` also changed to return an optional instead of
expecting a $default. But also completely changed in what it returns.
It does not return the number anymore, it just uses the $key for comparison
and returns the original element. So `$f_key` is just a selector
for comparison.