Option implementation
A comparison between two different Option implementations.

The first implementation uses an Array and uses the idea of zero or
one element to figure out if a value is present or not.

the second implementation directly uses a ref to a scalar and uses undef
to represent the none value.

Result:
The scalar implementation is marginally faster and has lower memory
footprint. But the gains are not big enough at the moment to again change
the Option implementation. At the moment i stick with the array
implementation.