Changed: Array::push
Removed the return value again. I thought it was a good idea, could be
handy at some places but I changed my mind again. It somehow breaks
with the idea that mutation functions should return nothing. This is
a design i try to keep as much as possible.

It could be handy in the fold() example as `$array->push` then returns
the $array again, so a user don't have to return $array explicitly. But
it is not a good idea to make that easier.

It is good that it feels awkward todo because in that case the user
should use fold_mut() instead that can provide a better perormance.