Changed: Option
This change is a little bit bigger, but overall should have less impact
on current code. It is now more flexible in how you can use option
or return values that expects an option.

The first change is that now optional values cannot be nested anymore.
Before you had the ability to write stuff like

    my $opt = Some(Some([]));

this contained an option containing an option. Now that is not possible
anymore. When you pass an optional to Some() than this optional is now
returned as-is. This change allows you to write stuff like.

   my $opt = Some( function($x,$y) );

This allows function() to either return an optional, or use a normal
value/undef to indicate the absence of a value. So you as a caller
to function() don't need to care anymore what function() returns. You
always get an optional back.

This is comfortable for both. The one who writes function() as the one
who calls function().

This behaviour leads to changes that now function that should return
an optional are free to either return an optional or just a normal
value/undef.

Currently Array::unfold, Array::choose and Array::pick are re-written
to use that ability.

Option->extract() and Option->extract_array() are added to allow
this kind of magic happening (or provide just a speed improvement)
without that you need to re-write the logic yourself.