Seq: range: inline range()
Sometimes annoying that we don't have an inline feature like in F#.
So from 1 line I changed it to 37 lines of code. But the reason is
because of a potential higher performance. Not important for every
function, but because range() is used so often i think it is important
here to give a full inlined version.