Changed: Option->extract_array
Change is backwards compatible. It is now even more convenient to use.

Usually Option->extract_array() is used on the result of calling a function.
When this function is called, this function can return multiple values
as an optional value.

It can indicate a failure by either returning nothing, a single undef or
None().

Multiple values are always considered valid, a single option with Some
is considered valid and returned as a single value. Multiple values can
be returned by returning Some([...]).