Parser: Added: p_fail, p_choose, p_qty, p_repeat
p_map always maps a parser and transforms a value, but it cannot make
a decision anymore if a value is valid or not. Same as Array::map you
just map a value.

p_choose instead can not only map a value, but also can decide if that
value is valid or not. Same as Array::choose.

p_qty allows a parser to repeat a minimum,maximum amount of times.