Removed Carp::croak
I first added it because it happened quite often that i defined a regex
but forget to add () to extract something. I thought it makes sense
to always have at least one capture.

I was wrong. For example defining `p_match(qr/\s*/)` to match any
whitespace is a use-case for using a regex and would be annoying todo
with p_str() and we also are not interested in extracting whitespace.