Deleted p_is
Not needed anymore. Previously I didn't know about @{^CAPTURE} variable.
So i had one function that extracted the match $1 explictily and p_is
that didn't extract anything.

But because of @{^CAPTURE} this distinction isn't needed anymore. It just
extracts everything or nothing. So we just need to define a regex without
captures to not extract anything. Basically it just means using (?: )
compared to just ().

It forces the user to use non-capturing groups if someone don't wanna
extract anything. As a side-effect the regex becomes faster.