Copy of current $ctx
Instead of creating a whole new $ctx the $ctx is copied with pos changed.
This makes it a little bit slower because of the function call, but
at the moment this is more correct.

In the near future $ctx will contain more things than just the current
position, and in that case those values must be copied not thrown away.