precompile regex
Doesn't seems to have an effect but i maybe expected that. Perl usually
compiles regexes only once, but it has a caching system as far i know
and only keeps a certain amount in the cache. Here i am doing that
explicitly. For the Benchmarks i used i didn't noticed a difference, but
this can change with a big parser and many regexes.