Changed: Optional
Optionals now can hold multiple values instead of just one value. After
some testing, coding especially by writing the Parser i came across the
problem that Option->extract_array might be a performance bottleneck
in the future. Also that function was so complex it might not really
possible to inline this.

That an optional can hold multiple values might sense as Perl also
have a scalar and list context. So an optional with multiple values is
somehow like an optional in list context.

For this change to work only small changes needs to be done.

1. Now parenthesis must be added in some places `[Some 1, Some 2]` now
   must be written `[Some(1), Some(2)]`.
2. When an `$f_opt_array` was expected for example in `Array->unfold` now
   you must write `Some($x,$state)` instead of `Some([$x,$state])`.

This change makes a lot of stuff easier. Option->extract_array is not
needed anymore. Every function can be wrapped in an `Some( function() )`
call and it works no matter if the function returns or many arguments.

There is only `Option->extract` now.

Some more tests must be added for this behaviour.