Parser: two version of p_many
p_many now returns two versions of itself depending if one or many parsers
are passed. When p_many is called with only one parser than the whole
for loop that emulates the p_and call is eliminated what results in
higher performance when p_many is called with just one parser.