Type: Changed Err messages
Changed the Error messages. In the future the Err case is very highly
replaced by a data-structure that contains more information on
what failed that then can be turned into a error string. But at the
moment this is fine.

Changed the tests to just use t_valid and don't check for error strings
that are returned.