Signature: Added sigt
`sigt` considers every function just as a single input and output
function. The whole input is considered as a "Tuple", because
of this the 't' is added to `sig`.

The whole input is wrapped into an array and then completely passed
and type-checked against the $in_type. This actually makes the whole
type-checking a lot more easier and more powerful. It's very easy
to now describe multiple different signatures and other kind of flexible
structures.

Actually `sig` already also works this way internally and creates a
t_tuple before checking. `sigt` in that sense just does less compared
to `sig`.