Type: Replaced t_str_eq -> t_enum
Anyway makes less sense to directly compare a string in a type-checker.
Even if theoretically a type could be determined this way when a
hash key is used a type field. But i guess better is to test
the structure itself.

But it makes sense to allow different string as enums, because Perl
otherwise has no enum implementation.