Type: internal data-structure
The Type system uses it's own internal data-structure that is basically
just a copy of the Result type. But i only need the basics. Reasons
for that is that i run into a deep-recursion problem when i tried
to add signature for results. And it makes sense as calling methods
on the result that then tries again to type-check that uses a data-type
to type check again ...

So here just a basic copy of result is used. Also should make it a little
bit faster because no blessing is done. And even when Result is added
with type-checking. The type system itself doesn't use it.

From the usage itself nothing changed. Running t_run still returns a
Result and a Result type is created. This is just an internal change.