Type: Changed internal representation
So far I used an Result-type. But for the valid case i just used `Ok 1`.
I often thought about that the Ok case could have some additional
info, but throw all of them away. It would make the type-system
"complex" in the sense that it then does more than just checking a type.

So only thing i really need was a value for an error-case or everything
is valid. So i am just using `undef` for valid case. I expect that
a check basically returns an error or not.

Currently it is just a string, but still can be anything later when i
wish to change this.

One reason for change was performance, but it doesn't really change much.

When a value is valid then the whole code and functions must be run, but
all of them shared anyway a single `Ok 1` that was created only once.

When a type-check fails it now can be a little bit faster. But wouldn't
anyway not matter much. It anyway bails out on the first error and doesn't
check all. So performance did not change much. But i consider code a little
bit cleaner now.

This change is internal only and should not change anything.