Added: t_as_hash
This considers an array as a hash, so you then can describe deeper
tests as a hash. This is useful in function arguments. As an
example see signature for Option::match where this is used to ensure
that someone passes arguments with Some,None.

Also because the signature check can do this now completely. I removed
the Carp::croak calls from the match() function.