Changed Exporting
Put Array, Hash, ... in the module they are written. So
`Sq::Collections::Array` is also in the package `Sq::Collections::Array`
instead of `Array`. This satisfies the CPAN Upload stuff. But the
functions are all now exported into `Array` when the module is loaded.
This is only done once, so loading the module multiple times does
not redefine the functions again and again.

This is done for all Collection modules.