Corrected Sq::Sig::Parser
Also fixed Sq::Reflect::get_func. Somehow it always returned a code-ref
even for not defined functions. This is now fixed. Because of this
i correctly realized that I used the wrong package name in this
signature file.

Still there is a problem with this signature, because the signature is
usually loaded after Sq::Parser is loaded and this modules imports
function into the calling namespace that at this point are not added
with type-checking.