Changed Queue
The normal Perl Array already does a good Job for using as a Queue.
So I replaced the implementation with just that. Actually I also
could remove Queue completely. But I think it is a good idea to
use a Queue when you want to have a Queue, with limited features
and it's own API instead of using an Array. This avoids confusion.