Seq: Some refactoring
Before skip($amount) defenitely always looped $amount time, even when
the iterator behind it was already exhausted. This changed, but
caustes more stress to the initial skipping.

Also it changed that the skipping is already done at the creation of
the iterator. Not in the code that just shouldreturn the next element.