Update POD to reflect changing error messages
That's why i want to have a better way of writing documentation. I need
a way to write code. And the output of the code should be inserted.
So there is a generation that creates the documentation.

There are sure systems that can do that. But i don't want to use
Latex and learn again another programming language. Also the syntax
is horrible.

Maybe something similar that was available in F#. There you could
have written an executable binary and put comments into it. The comments
was markdown and processing the file created Markdown. It also then
could put type information in the generated output.

It worked "okay". Had some problems with refering output and correctly
embedding pieces of code.

But would also be cool to have a general solution, not just one that
works only for Perl.