Test: Changed ok()
ok() is usually designed to just take a boolean value to indicate if
something succeeds or not. It shouldn't be used for any other values.

But i didn't checked for that. I just made a simple `if ( $bool )` check.

During testing i came around a problem that i didn't knew why a test passed
, and it should failed instead. The problem was that I passed an
Result type to `ok()` and this was considered a true value for `ok()`.
So the test passed. Those are sometimes the reasons why i hate
dynamic-typing.

So i searched maybe 2 hours to find an non existing bug, until i realized
that i passes the wrong thing to `ok()`. So annoying! So now i added
a type-check that for everything else than 0/1 the test-suite aborts.

So in the future this kind of "error" will not happen again.