Added prototype to sq
This allows to omit some () in some cases. As sq anyway just expects
a single value and this will not change. This is a good change.
Didn't do this before because it throwed a warning/error when prototype
was called on recursive subroutine.

Function prototype must be pre-declared than this works.