Different sq implementations
Function that operates on many arguments are faster. What makes sense
as the function then can internally loop through all elements, instead
that a function is called for every single element. This is even more
seen with sq3 that basically unfolds two loops.

On the other hand it is just around 4%-10% at maximum and then i
introduce a list context function again. Would like to avoid that.

Still interesting to know.