Refactor
I guess this is "slower" as before. But for printing a table that
doesn't matter much. It's at least better understandable what is going
on. At least in my opinion.

First with fill_blanks i create an Array of Array that is of same size
and fill empty spaces with 0. Then i zip all of the inner arrays. ziping
then combines all first elements, then all index 1 elements, then all
index 2 elements and so on.

Then i just get the maximum of those inner arrays. So it's a little bit
like a transpose and calling max on the inner arrays.