New: global 'static' function
This allows to create 'static' functions. A static function in C# for
example is a function that is directly called on a class.

    Class.whatever()

In Perl this is

    Class->whatever()

But in Perl an extra argument. The string `Class` is provided so someone
can use that to bless an object. In a static function this parameter
is not needed. Technically someone also could call.

    Class::whatever()

then the extra argument is not passed. Anyway this one extra argument
is sometimes annoying. For example is_prime is usually called this way.

    Sq->math->is_prime($x)

Here it is for convenient and it allows to build something like namespaces
on `Sq`. but it only works with `->`. Here calling `is_prime` will get
an additional `Sq::Math` passed as an extra argument.

The annoying part is that you cannot do.

    my $func = \&Sq::Math::is_prime;

Or, let's say, you could do. But then you get a sub-ref that also expects
that empty first argument that has no purpose. You would need to call

    $func->(undef, 100);

But if we could omit the first argument we actually could do.

    $seq->keep($func)
    $seq->keep(\&Sq::Math::is_prime);

in some way the `static` function allows you to write a function without
that extra leading class variable. But the function is installed in a
way that it will execute the underlying function when you pass more
than one argument.

Otherwise without any argument it directly returns the subroutine
reference instead of executing the function. This allows to write.

   $seq->keep(Sq->math->is_prime)

instead of

   $seq->keep(sub($x) { Sq->math->is_prime($x) })

the good part about it. The convenient way will also be the fastest
way (CPU performance) with the most less overhead.

So like a win-win situation.