Changed: Array::split signature
the built-in `split` in Perl takes a regex. It is good practice in Perl to
always pass a regex to split and not a string. For example `split '.', @array`
probably does not what you want.

It does not split something on a dot. It splits something on the first
character. as this is what `.` means in a regex. To avoid this common
mistakes where someone provides a string without realizing the string is
interpreted as a regex, now you must pass a regex, otherwise it is a
failure.

Run into the same problem even with knowing. So now i change that.