Changed: Array::init
Before this change Array::init could return multiple values, and
all of those values became part as a flattened Array. But this behaviour
is completely different to Seq::init.

I also wouldn't change Seq::init to this kind of behaviour. Seq::init
just expects one value and it aborts on an undef, that's how Seq works,
and now Array::init works the same.

In some tests i used Array::init to create a flattened array that then
gets converted to a Hash for example. Either use map{} for that
old behaviour.Or in the case of creating a Hash use Hash::init directly.
Don't create an Array and convert it to a Hash. That's how Perl does
it, but in Sq there is Hash::init for that.

Otherwise return an Array from Array::init and then flatten/merge it.
That's also how it would work on a Seq.