Refactored Signature
This can potential be a little bit faster and consuming less memory.
When you write `t_hash` then it is a function that creates a single
anonymous function that checks for a hash. So calling it ~60 times
also creates 60 of those functions. But as those functions are pure
they can be easily shared. Only one is needed.