doc: Added Sq::Fs | Change: Sq->fs->read_bytes
Added documentation for the Sq::Fs Module. While writing documentation
i changed the order of the arguments of the `read_bytes` function to
be the same as `read_raw`. Both expect the size to be read as first
argument. On top you now also can specify a @path instead of just
a single string.