Added: Array::contains
Consider that this scans an array. This is always slower than building
a hash and checking for the existence of a key. But a Hash-key only
can be a string. So sometimes it's hard to basically write a hashing
function that can turn any value into a string to be used.

contains() on the other hand just uses equal() on every value. So it
does the equality in Sq to check if a value is inside an Array.

You also can pass a Regexp for the contains() check.

When you not only want to check if an element is inside the Array but
also want to get the value than use `Array::find` or `Array::pick`