Added some signatures
But also removed signatures for dump/dumpw on Array/Hash/Option.
They anyway didn't made any sense and do not provide anything useful.

Dumping allows any value, so there is not really anything to check here.
Maybe the methods get removed in the future as dump() or dumpw() is
available as a function anyway.