Language feature 'static'
I will create a Manual Doc for that one. But `static` is for creating
static functions. When they are called without any arguments, the
function itself is returned. Instead of being executed.

On top the `static` is built into the rest of the system, especially
with the signature in mind. When adding a signature for a static
function it doesn't see it's first argument, only those that are
of interest.

Also the type-check is avoided for a static function when no argument
is passed. This is important otherwise `static` wouldn't work.