Changed: t_of, t_tuplev
When writing something new i encountered a problem with loading.
I still not fully understand why the way it was before didn't
worked correctly. But i fixed it by changing the loading of Sq::Type
like i had it before. I don't load `Sq`, manually import the functions
i need and that's it.

# t_tuplev
t_tuplev is now changed that the vararg check that was defined is always
executed. Before it only was executed when there was at least one element.
But in practice this seems weird. For example writing.

t_tuplev(t_str, t_array(t_min 1))

will not type-check that the vararags has at least one element and an
empty vararags also would be fine.

# t_of
But i used/test this behaviour. I allowed it to also be able to write
a fixed-size of optional arguments. Before you could write

t_tuplev(t_str, t_int, t_tuple(t_str, t_int))

So it was either two arguments, or four. But actually this doesn't
make sense. A tuple is the idea of being an array with fixed defined
size, so that this is suddenly optional seems not consistent.

Now it is changed, and writing the above would mean you must pass
three arguments. $str, $int [$str, $int]

The third being an array of exactly two values again.

The ability to allow zero or more elements of something is to use t_array.
But the problem was in combination with t_of. You only could specify
that all elements must be of the same type. t_of only supported
a single type.

This is now changed. Now you can specify multiple types.

t_array(t_of(t_str, t_int))

this now means it must be an array containing str,int one after another
either zero or many times. It then must be a multiple of two elements.
Writing three elements means the array must be a multiple of three
and so on.

A signature for writing 2 and then followed by str,int is now written

t_tuplev(t_str, t_int, t_array(t_of(t_str, t_int)))

Those data-structures would fullfill the above type-check (just see the
tests)

["foo", 1]
["foo", 1, "bar", 2]

an exact length of the array can be restricted with t_min, t_max or t_length.

Using t_of() with multiple values on a hash means that the value
can be of any defined types.

t_hash(t_of(t_str, t_array))

The above means a hash with values that are either strings or arrays.