Added: array/hash
global functions for creating blessed array/hash. Why another way?
Currently you can do `sq([1,2,3])` or `Array->new(1,2,3)`. But both ways
are sometimes not as fine as just writing `array(1,2,3)`.

The greatest benefits are that you just can pass those functions as a
lambda in certain cases to create an array when multiple varargs are
passed to a lambda.

Look at the tests.