Changed: Array: map2/map3
Those functions now changed how they work with arrays of unequal size.
Before it looped the maximum of the smallest array. So when you passed
an array with 3 and another with 5 elements then it made 3 iteration
at maximum.

Now it runs with 5 iterations. The last element of the smallest array
is just repeated.

Consider that it can have weird effects when you mutate elements. Only
properly works when you return new things. But that is anyway the default
way how nearly every function in Sq work.

When you don't want this behaviour and only need to run just the
smallest element, or need other values instead of repeating the last one,
then consider using fill() or take() on your arrays before.