Added: Fs: write_text()
write_text() has three modes, and it automatically does an encoding
to UTF-8. But only when the string you pass it is correctly in Perl's
internal unicode format (Perl uses UTF-8). Otherwise no encoding is done
and it is assumed a user already did an encode().

This way it will also work correctly when you already have an UTF-8
encoded string but unicode flag is not active, because you didn't
decode() it, did not `use utf8` or other kind of things.

# write_text($file, $string)

Just considers $string as the whole file-content and write it to a file.

# write_text($file, $array_of_strings)

When you pass it an array of strings, each string is considered a line.
An additional newline is automatically added at the end of every line.

# write_text($file, $seq_of_strings)

Writes a sequence of strings to a file, also here a newline is automatically
added.