Change: Fs: open_text()
open_text() is slightly changed that it now automatically chomp()
newlines at the end of every line. This is done so open_text()
and read_text() are isomorph.

Isomorph means the following. Consider you have a decimal number, convert
it to a hex, and then convert that hex back to a decimal. When the result
of this back-and-forth transformation is the same. Then two functions
are isomorph.

write_text() doesn't assume that you must put a newline at the end of
every string when you pass it an array of strings. So when you
have written an array of string, and then open a file and compare it,
those two data-structures wouldn't be the same.

The original wouldn't have newlines at the end of the strings, while
reading the file back into memory had the newlines. This is now changed.

You can write an array of strings. and reading/comparing it with the
written file is now the same.

From a practical point nothing really changes much. You anway get a
sequence of lines, so having an newline at the end of every string
is pretty much useless information. From a practical point chomp()ing
lines is anyway what you must often do when reading text-files.

This also might work fine in the future when you use some Sq->fmt functions
to otherwise print data again, then you very often want the newlines
removed to correctly work.