Fix: Found error while adding/changing signatures
Before the enum just wasn't right. Correct opening/closing was wrong.
So the script tag was basically never executed and script was just
considered a normal tag. While changing to `eq` i realized this error.

Then the next error was that it didn't returned [HTML =>]. And sure
must be correct, i need to return [HTML => ""] not [script => ""].