Load Path::Tiny
because the whole Sq::Fs is now lazy loaded, i don't need to lazy
load Path::Tiny in every function anymore. When no Sq->fs function
is called Path::Tiny is not loaded. Because this function is anyway
nearly used in every function i directly load it at the start.

This is different to Digest::SHA or PerlIO::gzip that i only load
when the function that uses those modules is loaded.