slice: Added TODO
Those functions now still needs some work. I need to make them compatible
so they work the same. For an Array we can provide negative indexes
also for a slice, but for a sequence this is a bad idea. To convert
a negative index to a real index i must iterate the whole sequence
once, count them and so on. Maybe cache everything. This isn't a good idea.

So i decided to drop the feature of supporting negative indexes. It's
important that Seq/Array both work the same, so Array now must be updated.

If someone want slices on array with negative indexes than someone just
can use Perl's built in slice syntax.