Equality: Added Path::Tiny
When are paths considered equal? Is a relative path to a file the
same as an absolute file to the same file, or not?

Let's assume the answer is yes, then I would need to make every
path absolute before comparision. But, not every path must exists and can
be made absolute. Paths itself can be purely 'virtual/made-up'.

I just decide to stringify and that's it. If needed someone must make
the paths absolute themself.