Improved nok()
ok() and nok() are written that they only expect 0,1 as valid values.
It does it this way so people who are writting "boolean functions"
really and only use 0 and 1 as boolean values, no other value.

But the function needs proper error-messages when wrong things are passed
to it. Added the "undef" value for eplicit checking.