Improved Dumping of Discriminated Union
When color dumping is active, what is the default, then cases are
printed with a unique Color that should identify them as a case,
otherwise without color the whole $union->case() is printed like before.

This helps greatly in reading a dumped data-structure build out
of DU. The dump is still valid code when "union->install()" was
made and all the cases are installed as functions.