Added TODO
with_default() maybe should be a lot more intelligent when it should
overwrites key/values in a Hash. This is a lot more of what i actually
how i want it to work. For example i have.

$hash->with_default(
    balance => 0,
    comment => ""
);

and "balance" stays an empty string "". But i assume it should be 0.
Same goes for other values. When i define a default empty Array for a key,
then i should be able to work with it as an array, and it should not be
an String, Number, Hash or whatever.