Changed: dump()
dump() is changed to return is input value as-is. In normal code so far
this will not change anything. But dump() is added as a method to various
packages like Array, Seq, Hash, Option, Result and so on. Now you can
call dump() as a method that dumps the data-structure at this point
and just call another method on it.