Updated CSV Example
Noe uses combine(). Still not exactly what i want. cartesian() returns
an array of hashes. not a single one. Theoretically in SQL you get a
single line represented as a single hash in Perl. But combining all
entries into a single hash in cartesian() is not good because it
can potentially overwrite a lot of keys. That's why an array of hashes
is choosen.

The choose() now does the WHERE and renaming in one operation.