Added: Array::dispatch, changed dispatch()
Changed dispatch() to take a hash-ref instead of multiple arguments. With
multiple arguments a hash-ref was built on every invocation of dispatch().
Inlcuding all the sub-refs. By taking a hash-ref you can built the
hash in advance. For example outside a loop, and then just pass it for
better performance. While it is theoretically possible to optimize, still
didn't know if it is so helpful, because very often we need to access
variables at a specific scope, and then we cannot put the hash-ref outside
a loop.

So i also implemented dispatch() by passing just one single argument. Then
it creates a function that expects the value to dispatch and any additional
values. So this created function can be used in a for-loop.

But i guess that looping over values and doing a dispatch is such a common
operation that i just added Array::dispatch for it.